Grab the pocket data from the apricorn.

+++ include/apricorn_tree.h
@@ -4,13 +4,13 @@
 #include "constants/apricorn_tree.h"
 #include "constants/items.h"
 
-bool8 IsApricornTreePicked(u8 id);
-void SetApricornTreePicked(u8 id);
+bool8 IsApricornTreePicked(u32 id);
+void SetApricornTreePicked(u32 id);
 
 void DailyResetApricornTrees(void);
 void ObjectEventInteractionGetApricornTreeData(void);
 void ObjectEventInteractionPickApricornTree(void);
-u8 GetApricornTypeByApricornTreeId(u8 id);
-u8 GetApricornCountByApricornTreeId(u8 id);
+u8 GetApricornTypeByApricornTreeId(u32 id);
+u8 GetApricornCountByApricornTreeId(u32 id);
 
 #endif //GUARD_APRICORN_TREE_H
